fix: рефакторит компоненты
