+++ src/hooks/store/index.ts
@@ -1,0 +1,2 @@
+import {useAppSelector, useActionCreators} from './store';
+export {useAppSelector, useActionCreators};
