+++ src/hooks/use-auth/index.ts
@@ -1,0 +1,2 @@
+import {useAuth} from './use-auth';
+export default useAuth;
