+++ src/hooks/use-auth/use-auth.ts
@@ -1,6 +1,6 @@
-import {useAppSelector} from './store';
-import {userSliceSelectors} from '../store/slices/user';
-import {AuthorizationStatus} from '../const';
+import {useAppSelector} from '../store';
+import {userSliceSelectors} from '../../store/slices/user';
+import {AuthorizationStatus} from '../../const';
 
 export function useAuth() {
   const authStatus = useAppSelector(userSliceSelectors.authorizationStatus);
