+++ src/hooks/use-form/index.ts
@@ -1,0 +1,2 @@
+import useForm from './use-form';
+export default useForm;
