+++ src/hooks/use-layout/index.ts
@@ -1,0 +1,2 @@
+import useLayout from './use-layout';
+export default useLayout;
