+++ src/hooks/use-map/index.ts
@@ -1,0 +1,2 @@
+import useMap from './use-map';
+export default useMap;
