+++ src/services/adapters/index.ts
@@ -1,0 +1,2 @@
+import {adaptOfferPreviewToApp, adaptOfferToApp, adaptOfferPreviewsToApp, adaptReviewsToApp} from './adapters';
+export {adaptOfferPreviewToApp, adaptOfferToApp, adaptOfferPreviewsToApp, adaptReviewsToApp};
