+++ src/services/api/index.ts
@@ -1,0 +1,2 @@
+import {createAPI} from './api';
+export {createAPI};
