+++ src/services/token/index.ts
@@ -1,0 +1,2 @@
+import {getToken, saveToken, deleteToken} from './token';
+export {getToken, saveToken, deleteToken};
